requirements.txt
absl-py==0.15.0
astunparse==1.6.3
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
beautifulsoup4==4.9.0
bs4==0.0.1
cachetools==4.2.4
certifi==2021.10.8
chardet==3.0.4
clang==5.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
cycler==0.11.0
debugpy @ file:///D:/bld/debugpy_1636043421411/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1631346842025/work
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
flatbuffers==1.12
fonttools==4.28.3
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.1.0
idna==2.10
importlib-metadata==4.8.2
ipykernel @ file:///C:/ci/ipykernel_1633545574250/work/dist/ipykernel-6.4.1-py3-none-any.whl
ipython @ file:///D:/bld/ipython_1638470483769/work
ipython-genutils==0.2.0
jedi @ file:///D:/bld/jedi_1637175281472/work
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1637611911738/work
jupyter-core @ file:///D:/bld/jupyter_core_1636814449073/work
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.6
matplotlib==3.5.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1631080358261/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1638419302549/work
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
pandas==1.0.5
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==8.4.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1639065841292/work
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1629119114968/work
pyparsing==3.0.6
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2021.3
pywin32==302
pyzmq @ file:///D:/bld/pyzmq_1635877383625/work
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.7.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.7.0
tensorflow-gpu==2.6.0
termcolor==1.1.0
tornado @ file:///D:/bld/tornado_1635819758615/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1635260543454/work
typing-extensions==3.7.4.3
urllib3==1.25.11
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
Werkzeug==2.0.2
wincertstore==0.2
wrapt==1.12.1
zipp==3.6.0
